refactor: cli sub package; top down definitions
