+++ cmd/gateway/root.go
@@ -6,3 +6,7 @@ var Cmd = &cobra.Command{
 	Use:   "gateway",
 	Short: "Manage a Storacha Network gateway",
 }
+
+func init() {
+	Cmd.AddCommand(serveCmd)
+}
