+++ cmd/upload/root.go
@@ -1,4 +1,4 @@
-package cmd
+package upload
 
 import (
 	"errors"
@@ -13,23 +13,33 @@ import (
 	"github.com/storacha/go-ucanto/did"
 
 	"github.com/storacha/guppy/cmd/internal/upload/ui"
+	"github.com/storacha/guppy/cmd/upload/source"
 	"github.com/storacha/guppy/internal/cmdutil"
 	"github.com/storacha/guppy/pkg/bus"
 	"github.com/storacha/guppy/pkg/config"
 	"github.com/storacha/guppy/pkg/preparation"
-	"github.com/storacha/guppy/pkg/preparation/spaces/model"
 	"github.com/storacha/guppy/pkg/preparation/sqlrepo"
 	"github.com/storacha/guppy/pkg/preparation/types"
 	uploadsmodel "github.com/storacha/guppy/pkg/preparation/uploads/model"
 )
 
-var uploadFlags struct {
+var log = logging.Logger("cmd/upload")
+
+var rootFlags struct {
 	all         bool
 	retry       bool
 	parallelism uint64
 }
 
-var uploadCmd = &cobra.Command{
+func init() {
+	Cmd.Flags().BoolVar(&rootFlags.all, "all", false, "Upload all sources (even if arguments are provided)")
+	Cmd.Flags().BoolVar(&rootFlags.retry, "retry", false, "Auto-retry failed uploads")
+	Cmd.Flags().Uint64Var(&rootFlags.parallelism, "parallelism", 6, "Number of parallel shard uploads to perform concurrently")
+
+	Cmd.AddCommand(source.Cmd)
+}
+
+var Cmd = &cobra.Command{
 	Use:     "upload <space> [source-path-or-name...]",
 	Aliases: []string{"up"},
 	Short:   "Upload data to a Storacha space",
@@ -44,7 +54,7 @@ var uploadCmd = &cobra.Command{
 		space := args[0]
 		if space == "" {
 			cmd.SilenceUsage = false
-			return errors.New("space cannot be empty")
+			return fmt.Errorf("space required")
 		}
 		spaceDID, err := did.Parse(space)
 		if err != nil {
@@ -80,7 +90,7 @@ var uploadCmd = &cobra.Command{
 		// defer repo.Close()
 
 		api := preparation.NewAPI(repo, cmdutil.MustGetClient(cfg.Repo.Dir),
-			preparation.WithBlobUploadParallelism(int(uploadFlags.parallelism)),
+			preparation.WithBlobUploadParallelism(int(rootFlags.parallelism)),
 			preparation.WithEventBus(eb),
 		)
 		allUploads, err := api.FindOrCreateUploads(ctx, spaceDID)
@@ -89,13 +99,15 @@ var uploadCmd = &cobra.Command{
 		}
 
 		if len(allUploads) == 0 {
-			fmt.Printf("No sources found for space. Add a source first with:\n\n$ %s %s <path>\n\n", uploadSourcesAddCmd.CommandPath(), spaceDID)
+			fmt.Printf("No sources found for space. Add a source first with:\n\n$ %s %s <path>\n\n",
+				source.AddCmd.CommandPath(),
+				spaceDID)
 			return cmdutil.NewHandledCliError(fmt.Errorf("no uploads found for space %s", spaceDID))
 		}
 
 		var uploadsToRun []*uploadsmodel.Upload
 
-		if len(requestedSources) == 0 || uploadFlags.all {
+		if len(requestedSources) == 0 || rootFlags.all {
 			uploadsToRun = allUploads
 		} else {
 			reqMap := make(map[string]bool)
@@ -128,7 +140,7 @@ var uploadCmd = &cobra.Command{
 		}
 
 		if useUI {
-			return ui.RunUploadUI(ctx, repo, api, uploadsToRun, uploadFlags.retry, eb)
+			return ui.RunUploadUI(ctx, repo, api, uploadsToRun, rootFlags.retry, eb)
 		}
 		// UI disabled, log at info level
 		logging.SetAllLoggers(logging.LevelInfo)
@@ -165,7 +177,7 @@ var uploadCmd = &cobra.Command{
 				var re types.RetriableError
 				if errors.As(err, &re) {
 					lastErr = err
-					if uploadFlags.retry {
+					if rootFlags.retry {
 						log.Warnw("Retriable upload error encountered, retrying", "upload", u.ID(), "attempt", attempt,
 							"err", err)
 						continue
@@ -213,174 +225,4 @@ var uploadCmd = &cobra.Command{
 		}
 		return nil
 	},
-}
-
-func init() {
-	uploadCmd.Flags().BoolVar(&uploadFlags.all, "all", false, "Upload all sources (even if arguments are provided)")
-	uploadCmd.Flags().BoolVar(&uploadFlags.retry, "retry", false, "Auto-retry failed uploads")
-	uploadCmd.Flags().Uint64Var(&uploadFlags.parallelism, "parallelism", 6, "Number of parallel shard uploads to perform concurrently")
-}
-
-var uploadSourceCmd = &cobra.Command{
-	Use: "source",
-}
-
-func init() {
-	uploadCmd.AddCommand(uploadSourceCmd)
-}
-
-var uploadSourcesAddFlags struct {
-	shardSize string
-	name      string
-}
-
-var uploadSourcesAddCmd = &cobra.Command{
-	Use:   "add <space> <path>",
-	Short: "Add a source to a space",
-	Long: wordwrap.WrapString(
-		"Adds a source to a space. A source is currently a path on the local "+
-			"filesystem, but this may be expanded in the future to include other "+
-			"types of data sources. `upload` will upload data from all sources "+
-			"associated with a space. Sources are associated with the space locally "+
-			"for future local upload commands; no association is made remotely.",
-		80),
-	Args: cobra.ExactArgs(2),
-
-	RunE: func(cmd *cobra.Command, args []string) error {
-		ctx := cmd.Context()
-
-		cfg, err := config.Load[config.Config]()
-		if err != nil {
-			return err
-		}
-		repo, err := preparation.OpenRepo(ctx, cfg.Repo.DatabasePath())
-		if err != nil {
-			return err
-		}
-		defer repo.Close()
-
-		space := cmd.Flags().Arg(0)
-		if space == "" {
-			cmd.SilenceUsage = false
-			return errors.New("space cannot be empty")
-		}
-
-		path := cmd.Flags().Arg(1)
-		if path == "" {
-			cmd.SilenceUsage = false
-			return errors.New("path cannot be empty")
-		}
-
-		path, err = filepath.Abs(path)
-		if err != nil {
-			return fmt.Errorf("resolving absolute path: %w", err)
-		}
-
-		spaceDID, err := did.Parse(space)
-		if err != nil {
-			cmd.SilenceUsage = false
-			return fmt.Errorf("parsing space DID: %w", err)
-		}
-
-		api := preparation.NewAPI(repo, cmdutil.MustGetClient(cfg.Repo.Dir))
-
-		// Parse shard size if provided
-		var spaceOptions []model.SpaceOption
-		if uploadSourcesAddFlags.shardSize != "" {
-			shardSize, err := cmdutil.ParseSize(uploadSourcesAddFlags.shardSize)
-			if err != nil {
-				return fmt.Errorf("parsing shard size: %w", err)
-			}
-			spaceOptions = append(spaceOptions, model.WithShardSize(shardSize))
-		}
-
-		name := path
-		if uploadSourcesAddFlags.name != "" {
-			name = uploadSourcesAddFlags.name
-		}
-
-		_, err = api.FindOrCreateSpace(ctx, spaceDID, spaceDID.String(), spaceOptions...)
-		if err != nil {
-			return fmt.Errorf("command failed to create space: %w", err)
-		}
-
-		source, err := api.CreateSource(ctx, name, path)
-		if err != nil {
-			return fmt.Errorf("command failed to create source: %w", err)
-		}
-
-		err = repo.AddSourceToSpace(ctx, spaceDID, source.ID())
-		if err != nil {
-			return fmt.Errorf("command failed to add source to space: %w", err)
-		}
-
-		return nil
-	},
-}
-
-func init() {
-	uploadSourceCmd.AddCommand(uploadSourcesAddCmd)
-	uploadSourcesAddCmd.Flags().StringVar(&uploadSourcesAddFlags.shardSize, "shard-size", "", "Shard size for the space (e.g., 1024, 512B, 100K, 50M, 2G)")
-	uploadSourcesAddCmd.Flags().StringVar(&uploadSourcesAddFlags.name, "name", "", "Name (alias) for the source")
-}
-
-var uploadSourcesListCmd = &cobra.Command{
-	Use:     "list <space>",
-	Aliases: []string{"ls"},
-	Short:   "List sources added to a space",
-	Long:    `Lists the sources added to a space.`,
-	Args:    cobra.ExactArgs(1),
-
-	RunE: func(cmd *cobra.Command, args []string) error {
-		ctx := cmd.Context()
-
-		cfg, err := config.Load[config.Config]()
-		if err != nil {
-			return err
-		}
-		repo, err := preparation.OpenRepo(ctx, cfg.Repo.DatabasePath())
-		if err != nil {
-			return err
-		}
-		defer repo.Close()
-
-		space := cmd.Flags().Arg(0)
-		if space == "" {
-			cmd.SilenceUsage = false
-			return errors.New("space cannot be empty")
-		}
-
-		spaceDID, err := did.Parse(space)
-		if err != nil {
-			cmd.SilenceUsage = false
-			return fmt.Errorf("parsing space DID: %w", err)
-		}
-
-		sourceIDs, err := repo.ListSpaceSources(ctx, spaceDID)
-		if err != nil {
-			return err
-		}
-
-		fmt.Printf("Sources for space %s:\n", spaceDID)
-		for _, sourceID := range sourceIDs {
-			source, err := repo.GetSourceByID(ctx, sourceID)
-			if err != nil {
-				return fmt.Errorf("failed to get source by ID %s: %w", sourceID, err)
-			}
-			if source.Name() != source.Path() {
-				fmt.Printf("- %s: %s\n", source.Name(), source.Path())
-			} else {
-				fmt.Printf("- %s\n", source.Path())
-			}
-		}
-		if len(sourceIDs) == 0 {
-			fmt.Printf("No sources found for space %s. Add a source first with:\n\n$ %s %s <path>\n\n", spaceDID, uploadSourcesAddCmd.CommandPath(), spaceDID)
-		}
-
-		return nil
-	},
-}
-
-func init() {
-	uploadSourceCmd.AddCommand(uploadSourcesListCmd)
 }
